Make 3D pose estimation notebook standalone
