--- notebooks/3D-pose-estimation-webcam/engine/one_euro_filter.py
@@ -1,64 +1,0 @@
-"""
- Copyright (c) 2022 Intel Corporation
- Licensed under the Apache License, Version 2.0 (the "License");
- you may not use this file except in compliance with the License.
- You may obtain a copy of the License at
-      http://www.apache.org/licenses/LICENSE-2.0
- Unless required by applicable law or agreed to in writing, software
- distributed under the License is distributed on an "AS IS" BASIS,
- WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
- See the License for the specific language governing permissions and
- limitations under the License.
-"""
-
-import math
-
-
-def get_alpha(rate=30, cutoff=1):
-    tau = 1 / (2 * math.pi * cutoff)
-    te = 1 / rate
-    return 1 / (1 + tau / te)
-
-
-class LowPassFilter:
-    def __init__(self):
-        self.x_previous = None
-
-    def __call__(self, x, alpha=0.5):
-        if self.x_previous is None:
-            self.x_previous = x
-            return x
-        x_filtered = alpha * x + (1 - alpha) * self.x_previous
-        self.x_previous = x_filtered
-        return x_filtered
-
-
-class OneEuroFilter:
-    def __init__(self, freq=15, mincutoff=1, beta=1, dcutoff=1):
-        self.freq = freq
-        self.mincutoff = mincutoff
-        self.beta = beta
-        self.dcutoff = dcutoff
-        self.filter_x = LowPassFilter()
-        self.filter_dx = LowPassFilter()
-        self.x_previous = None
-        self.dx = None
-
-    def __call__(self, x):
-        if self.dx is None:
-            self.dx = 0
-        else:
-            self.dx = (x - self.x_previous) * self.freq
-        dx_smoothed = self.filter_dx(self.dx, get_alpha(self.freq, self.dcutoff))
-        cutoff = self.mincutoff + self.beta * abs(dx_smoothed)
-        x_filtered = self.filter_x(x, get_alpha(self.freq, cutoff))
-        self.x_previous = x
-        return x_filtered
-
-
-if __name__ == "__main__":
-    filter = OneEuroFilter(freq=15, beta=0.1)
-    for val in range(10):
-        x = val + (-1) ** (val % 2)
-        x_filtered = filter(x)
-        print(x_filtered, x)
